Remove easing and disperse interpolation
resolves #6
