+++ src/main/kotlin/Floats.kt
@@ -39,3 +39,31 @@ public inline fun startOfChunk(x: Float, size: Float, origin: Float = 0F): Float
 
 public inline fun indexOfChunk(x: Float, size: Float, origin: Float = 0F): Float =
     (startOfChunk(x, size, origin) - origin) / size
+
+// Interpolation
+
+public inline fun <T> cerp(from: T, to: T, by: Float, threshold: Float = 1F): T = if (by < threshold) from else to
+public inline fun lerp(from: Float, to: Float, by: Float): Float = from + (to - from) * by
+public inline fun normalize(x: Float, min: Float, max: Float): Float =
+    if (min == max) 0F else (x - min) / (max - min)
+
+public inline fun repeat(x: Float): Float = x.mod(1F)
+
+public inline fun reflect(x: Float): Float {
+    val dist = abs(x) % 2F
+    return if (dist < 1F) dist else 2F - dist
+}
+
+public inline fun repeat(x: Float, min: Float, max: Float): Float =
+    lerp(min, max, repeat(normalize(x, min, max)))
+
+public inline fun reflect(x: Float, min: Float, max: Float): Float =
+    lerp(min, max, reflect(normalize(x, min, max)))
+
+public fun interface InterpolatedF<T> {
+    public fun interpolate(b: T, t: Float): T
+}
+
+public fun interface InterpolatorF<T> {
+    public fun interpolate(a: T, b: T, t: Float): T
+}
